Environment Variables set

+++ controllers/template.controller.js
@@ -3,6 +3,11 @@ const Template = require('../models/template.model');
 
 exports.getTemplates = async (req,res) => {
     let shop = req.params.shop;
+    if (!shop) {
+        return res.status(422).json({
+          error: "Shop field is required in URL params ",
+        });
+    }
     let response={};
     try{
         let templates = await Template.find({shop});
@@ -16,15 +21,20 @@ exports.getTemplates = async (req,res) => {
 }
 
 exports.createTemplate = async (req,res) => {
-    let { title, description, design_json, shop } = req.body
+    let { title, design_json, shop } = req.body
+    if (!title || !design_json || !shop) {
+        return res.status(422).json({
+          error: "Mandatory fields in request body unprocessable",
+        });
+    }
     let response = {}
     try{
-        let template = new Template({title,description,shop,design_json});
+        let template = new Template({title,shop,design_json});
         await template.save()
         response = {created:true, template}
     }
     catch(err){
-        console.log("Err",err)
+        console.log("Error Creating Template : ",err)
         response={created:false,error:err}
     }
     return res.status(200).json(response)
@@ -33,6 +43,17 @@ exports.createTemplate = async (req,res) => {
 exports.updateTemplate = async (req,res) => {
     let templateId = req.params.id;
     let {title, design_json} = req.body
+    if (!templateId ) {
+        return res.status(422).json({
+          error: "Template Id is required in URL param",
+        });
+    }
+    if (!design_json ) {
+        return res.status(422).json({
+          error: "Mandatory fields in request body unprocessable : design_json*",
+        });
+    }
+
     let response = {updated:null};
     try{
         let updateData = {design_json};
@@ -51,6 +72,11 @@ exports.updateTemplate = async (req,res) => {
 
 exports.deleteTemplate = async (req,res) => {
     let templateId = req.params.id;
+    if (!templateId) {
+        return res.status(422).json({
+          error: "Template Id required in URL param",
+        });
+    }
     let response = {deleted:null};
     try{
         await Template.findOneAndDelete({
